Move Frame in a separate file
