+++ src/api.rs
@@ -10,6 +10,7 @@
 use arg_enum_proc_macro::ArgEnum;
 use bitstream_io::*;
 use crate::encoder::*;
+use crate::frame::Frame;
 use crate::metrics::calculate_frame_psnr;
 use crate::partition::*;
 use crate::rate::RCState;
