+++ src/rdo.rs
@@ -19,8 +19,8 @@ use crate::header::ReferenceMode;
 use crate::encode_block_a;
 use crate::encode_block_b;
 use crate::encode_block_with_modes;
-use crate::FrameInvariants;
-use crate::Frame;
+use crate::encoder::FrameInvariants;
+use crate::frame::Frame;
 use crate::luma_ac;
 use crate::me::*;
 use crate::motion_compensate;
