+++ src/scenechange/mod.rs
@@ -7,7 +7,7 @@
 // Media Patent License 1.0 was not distributed with this source code in the
 // PATENTS file, you can obtain it at www.aomedia.org/license/patent.
 
-use crate::encoder::Frame;
+use crate::frame::Frame;
 use crate::util::{CastFromPrimitive, Pixel};
 
 use std::sync::Arc;
