+++ src/tiling/tile_state.rs
@@ -11,6 +11,7 @@ use super::*;
 
 use crate::context::*;
 use crate::encoder::*;
+use crate::frame::*;
 use crate::plane::*;
 use crate::quantize::*;
 use crate::rdo::*;
